onUserImageChange function

+++ socialapp-functions/functions/index.js
@@ -63,9 +63,9 @@ exports.api = functions.region("europe-west1").https.onRequest(app);
 
 exports.createNotificationOnLike = functions.region("europe-west1").firestore.document('likes/{id}')
 .onCreate(snapshot=>{
-    db.doc(`/screams/${snapshot.data().screamId}`).get()
+    return db.doc(`/screams/${snapshot.data().screamId}`).get()
     .then(doc=>{
-        if(doc.exists){
+        if(doc.exists && doc.data().userHandle !== snapshot.data().userHandle){
             return db.doc(`/notifications/${snapshot.id}`).set({
                 createdAt : new Date().toISOString(),
                 recipient : doc.data().userHandle,
@@ -77,29 +77,23 @@ exports.createNotificationOnLike = functions.region("europe-west1").firestore.do
             })
         }
     })
-    .then(()=>{
-        return ;
-    })
     .catch((error=>{
         console.log(error);
-        return;
+       
     }))
 });
 
 exports.deleteNotificationOnUnlike = functions.region("europe-west1").firestore.document('likes/{id}')
 .onDelete(snapshot=>{
-    db.doc(`/screams/${snapshot.data().screamId}`).get()
+   return  db.doc(`/screams/${snapshot.data().screamId}`).get()
     .then(doc=>{
-        if(doc.exists){
+        if(doc.exists && doc.data().userHandle !== snapshot.data().userHandle){
             return db.doc(`/notifications/${snapshot.id}`).delete()
         }
     })
-    .then(()=>{
-        return ;
-    })
     .catch((error=>{
         console.log(error);
-        return;
+       
     }))
 });
 
@@ -109,9 +103,9 @@ exports.deleteNotificationOnUnlike = functions.region("europe-west1").firestore.
 
 exports.createNotificationOnComment = functions.region("europe-west1").firestore.document('comments/{id}')
 .onCreate(snapshot=>{
-    db.doc(`/screams/${snapshot.data().screamId}`).get()
+   return db.doc(`/screams/${snapshot.data().screamId}`).get()
     .then(doc=>{
-        if(doc.exists){
+        if(doc.exists && doc.data().userHandle !== snapshot.data().userHandle){
             return db.doc(`/notifications/${snapshot.id}`).set({
                 createdAt : new Date().toISOString(),
                 recipient : doc.data().userHandle,
@@ -123,11 +117,29 @@ exports.createNotificationOnComment = functions.region("europe-west1").firestore
             })
         }
     })
-    .then(()=>{
-        return ;
-    })
+  
     .catch((error=>{
         console.log(error);
-        return;
     }))
+})
+
+
+exports.onUserImageChange = functions.region("europe-west1").firestore.document('/users/{userId}')
+.onUpdate((change)=>{
+    console.log(change.before.data());
+    console.log(change.after.data());
+
+    if(change.before.data().imageUrl !== change.after.data().imageUrl){
+        let batch = db.batch();
+        return db.collection('screams').where('userHandle', '==', change.before.data().handle).get()
+    
+        .then((data)=>{
+            data.forEach(doc=>{
+                const scream = db.doc(`/screams/${doc.id}`);
+                batch.update(scream,{imageUrl : change.after.data().imageUrl})
+            })
+            return batch.commit();
+        })
+    }
+   
 })
